backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.6.0
python-dotenv>=1.0.1
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
crewai>=0.120.0
pydantic>=2.9.0
